3. 문의글, 문의답변, FAQ
1. 문의글 추가, 수정 (엔티티, 저장소, 컨트롤러, 서비스)
2. 문의답변 추가, 수정(저장소, 컨트롤러, 서비스)
3. FAQ 추가(컨트롤러, 서비스)

+++ abmarket/src/main/java/com/example/restcontroller/AnswerRestController3.java
@@ -5,8 +5,10 @@ import java.util.Map;
 
 import com.example.entity.AnswerEntity;
 import com.example.entity.InquireEntity;
+import com.example.entity.MemberEntity;
 import com.example.jwt.JwtUtil;
 import com.example.repository.AnswerRepository3;
+import com.example.repository.MemberRespository2;
 import com.example.service.AnswerService3;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +26,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class AnswerRestController3 {
 
     // 문의 답변은 댓글 형태로 설정(관리자가 입력)
-    // 관리자가 아닌 사람은 등록 수정 삭제 X
     
     // 토큰
     @Autowired JwtUtil jwtUtil;
@@ -33,11 +34,13 @@ public class AnswerRestController3 {
 
     @Autowired AnswerService3 anService3;
 
+    @Autowired MemberRespository2 memRepository2;
+
     
     // 답변 등록 (관리자 토큰)
     // 127.0.0.1:9090/ROOT/api/answer/insertone
     @RequestMapping(value = {"/insertone"},
-    method = {RequestMethod.POST},
+        method = {RequestMethod.POST},
         consumes = {MediaType.ALL_VALUE},
         produces = {MediaType.APPLICATION_JSON_VALUE}
     )
@@ -52,20 +55,29 @@ public class AnswerRestController3 {
             //토큰 필요함(토큰 추출)
             String userid = jwtUtil.extractUsername(token);
             System.out.println("RequestMapping username : " + userid);
+            
+            //관리자
+            MemberEntity mem = memRepository2.getById(userid);
+            // System.out.println(mem);
 
-            InquireEntity inq = new InquireEntity();
-            inq.setInqno(inqno);
-            System.out.println(inq.toString());
+            //토큰 = ADMIN
+            if(mem.getUrole().equals("ADMIN")) {
+                InquireEntity inq = new InquireEntity();
+                inq.setInqno(inqno);
+                System.out.println(inq.toString());
 
-            answerEntity.setInquire(inq);
-            System.out.println(answerEntity.toString());
+                answerEntity.setInquire(inq);
+                System.out.println(answerEntity.toString());
 
-            int ret = anService3.insertAnswer(answerEntity);
-            if(ret == 1) {
-                map.put("status", 200);
-            }
-            else {
-                map.put("status", 0);
+                int ret = anService3.insertAnswer(answerEntity);
+                if(ret == 1) {
+                    map.put("result", "등록완료");
+                    map.put("status", 200);
+                }
+                else {
+                    map.put("result", "관리자X");
+                    map.put("status", 0);
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -91,13 +103,18 @@ public class AnswerRestController3 {
             //토큰 필요함(토큰 추출)
             String username = jwtUtil.extractUsername(token);
             System.out.println("RequestMapping username : " + username);
-            
-            int ret = anService3.deleteOneAnswer(anno);
-            if(ret == 1) {
-                map.put("result", "삭제완료!");
-                map.put("status", 200);
+
+            MemberEntity mem = memRepository2.getById(username);
+
+            if(mem.getUrole().equals("ADMIN")) {
+                int ret = anService3.deleteOneAnswer(anno);
+                if(ret == 1) {
+                    map.put("result", "삭제완료!");
+                    map.put("status", 200);
+                }
             }
             else {
+                map.put("result", "관리자X");
                 map.put("status", 0);
             }
         } catch (Exception e) {
@@ -124,11 +141,15 @@ public class AnswerRestController3 {
             //토큰 필요함(토큰 추출)
             String userid = jwtUtil.extractUsername(token);
             System.out.println("RequestMapping username : " + userid);
-            
-            AnswerEntity answerEntity = anService3.selectOneAnswer(anno);
-            if(answerEntity != null) {
-                map.put("status", 200);
-                map.put("result", answerEntity);
+
+            MemberEntity mem = memRepository2.getById(userid);
+
+            if(mem.getUrole().equals("ADMIN")) {
+                AnswerEntity answerEntity = anService3.selectOneAnswer(anno);
+                if(answerEntity != null) {
+                    map.put("status", 200);
+                    map.put("result", answerEntity);
+                }
             }
             else {
                 map.put("status", 0);
@@ -158,22 +179,26 @@ public class AnswerRestController3 {
             String userid = jwtUtil.extractUsername(token);
             System.out.println("RequestMapping username : " + userid);
 
-            //답변 기존 데이터 불러오기
-            AnswerEntity result = anService3.selectOneAnswer(answeren.getAnno());
-            //System.out.println("기존==="+result.toString());
+            MemberEntity mem = memRepository2.getById(userid);
 
-            //답변 수정 (내용)
-            result.setAncontent(answeren.getAncontent());
-            //System.out.println("새로운==="+result.getAncontent());
-
-            //변경 후 저장
-            int ret = anService3.updateOneAnswer(result);
-            if(ret == 1) {
-                map.put("result", "수정완료!");
-                map.put("status", 200);
+            if(mem.getUrole().equals("ADMIN")) {
+                //답변 기존 데이터 불러오기
+                AnswerEntity result = anService3.selectOneAnswer(answeren.getAnno());
+                //System.out.println("기존==="+result.toString());
+    
+                //답변 수정 (내용)
+                result.setAncontent(answeren.getAncontent());
+                //System.out.println("새로운==="+result.getAncontent());
+    
+                //변경 후 저장
+                int ret = anService3.updateOneAnswer(result);
+                if(ret == 1) {
+                    map.put("result", "수정완료!");
+                    map.put("status", 200);
+                }
             }
             else {
-                map.put("result", "수정실패!");
+                map.put("result", "관리자X");
                 map.put("status", 0);
             }
         } catch (Exception e) {
